requirements.txt
albumentations==1.4.20
opencv-python==4.10.0.84
pyyaml==6.0.2
tqdm==4.66.5
numpy==1.26.4
fastapi==0.115.0
uvicorn==0.30.6
celery==5.4.0
redis==5.0.8
python-dotenv==1.0.1
requests==2.32.3
psycopg2-binary==2.9.9
python-multipart==0.0.20
